LayoutTests/fast/events/ios/meaningful-click-when-playing-media.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<style>
body, html {
    font-size: 18px;
    font-family: system-ui;
    width: 100%;
    height: 100%;
    margin: 0;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;
didPauseVideo = false;
didPlayVideo = false;

addEventListener("load", async () => {
    description("This test exercises the 'meaningful click' heuristic when playing or pausing a video, and requires WebKitTestRunner.");
    document.body.addEventListener("click", async () => {
        const video = document.querySelector("video");
        if (video.paused) {
            await video.play();
            didPlayVideo = true;
        } else {
            await video.pause();
            didPauseVideo = true;
        }
    });

    if (!window.testRunner)
        return;

    async function simulateTap(shouldPlay) {
        didHandleTap = false;
        didHandleTapAsMeaningfulClick = false;
        testRunner.installDidHandleTapCallback(wasMeaningful => {
            didHandleTap = true;
            didHandleTapAsMeaningfulClick = wasMeaningful;
        });

        await UIHelper.activateElement(document.body);
        await UIHelper.waitForDoubleTapDelay();
        await new Promise(resolve => shouldBecomeEqual(shouldPlay ? "didPlayVideo" : "didPauseVideo", "true", resolve));
        shouldBeTrue("didHandleTap");
        shouldBeTrue("didHandleTapAsMeaningfulClick");
        testRunner.clearTestRunnerCallbacks();
    }

    await simulateTap(true);
    await simulateTap(false);
    finishJSTest();
});
</script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
    <video loop src="../../images/resources/animated-red-green-blue.mp4"></video>
</body>
</html>
