LayoutTests/fast/events/ios/meaningful-click-when-playing-media-expected.txt

This test exercises the 'meaningful click' heuristic when playing or pausing a video, and requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS didPlayVideo became true
PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is true
PASS didPauseVideo became true
PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is true
PASS successfullyParsed is true

TEST COMPLETE
